requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
torch>=2.1.0
numpy>=1.26.0
pandas>=2.1.0
pyyaml>=6.0
pydantic>=2.5.0
scikit-learn>=1.3.0
scipy>=1.11.0
pytest>=7.4.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
influxdb>=5.3.0
influxdb-client>=1.38.0
